Support for different types of fruit, currently only one more.
Sent ripple effects through the entire code base. Mostly because
it became apparent that Game_get_address returning in the same
address it's called with is completely unmaintainable. Added
Game_get_address_IX to return it in IX instead, which makes
everything so much easier. Needless to say, it uncovered several
bugs.

The old routine is still around, but should be removed some day.